+++ noti-service/src/main/java/com/waither/notiservice/api/NotificationController.java
@@ -1,14 +1,12 @@
 package com.waither.notiservice.api;
 
 import com.waither.notiservice.api.response.NotificationResponse;
-import com.waither.notiservice.dto.LocationDto;
+import com.waither.notiservice.api.request.LocationDto;
+import com.waither.notiservice.global.response.ApiResponse;
 import com.waither.notiservice.service.NotificationService;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,31 +18,26 @@ public class NotificationController {
 
     private final NotificationService notificationService;
 
-    @Operation(summary = "get notification", description = "알림 목록 조회하기")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "OK",
-                    content = @Content(schema = @Schema(implementation = NotificationResponse.class))),
-            @ApiResponse(responseCode = "400", description = "BAD REQUEST"),
-            @ApiResponse(responseCode = "404", description = "NOT FOUND"),
-            @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
-    })
+    @Operation(summary = "Get notification", description = "알림 목록 조회하기")
     @GetMapping("")
-    public List<NotificationResponse> getNotifications(Long userId) {
-        return notificationService.getNotifications(userId);
+    public ApiResponse<?> getNotifications(Long userId) {
+        return ApiResponse.onSuccess(notificationService.getNotifications(userId));
+    }
+
+    @Operation(summary = "Delete notification", description = "알림 삭제하기")
+    @DeleteMapping("")
+    public ApiResponse<?> delteNotification(@RequestParam("id") String notificationId) {
+        notificationService.deleteNotification(notificationId);
+        return ApiResponse.onSuccess(HttpStatus.OK);
     }
 
     @Operation(summary = "Send Go Out Alarm", description = "외출 알림 전송하기")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "OK"),
-            @ApiResponse(responseCode = "400", description = "BAD REQUEST"),
-            @ApiResponse(responseCode = "404", description = "NOT FOUND"),
-            @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
-    })
     @PostMapping("/goOut")
     public void sendGoOutAlarm(Long userId) {
         notificationService.sendGoOutAlarm(userId);
     }
 
+    @Operation(summary = "Send Go Out Alarm", description = "외출 알림 전송하기")
     @PostMapping("/")
     public void checkCurrentAlarm(@RequestBody LocationDto locationDto) {
         notificationService.checkCurrentAlarm(locationDto);
