+++ noti-service/src/main/java/com/waither/notiservice/service/NotificationService.java
@@ -3,14 +3,15 @@ package com.waither.notiservice.service;
 import com.waither.notiservice.api.response.NotificationResponse;
 import com.waither.notiservice.domain.UserData;
 import com.waither.notiservice.domain.UserMedian;
-import com.waither.notiservice.dto.LocationDto;
+import com.waither.notiservice.api.request.LocationDto;
+import com.waither.notiservice.global.exception.CustomException;
+import com.waither.notiservice.global.response.ErrorCode;
 import com.waither.notiservice.repository.NotificationRepository;
 import com.waither.notiservice.repository.UserDataRepository;
 import com.waither.notiservice.repository.UserMedianRepository;
 import com.waither.notiservice.utils.TemperatureUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -33,44 +34,56 @@ public class NotificationService {
                 .stream().map(NotificationResponse::of).toList();
     }
 
+    public void deleteNotification(String notificationId) {
+        notificationRepository.deleteById(notificationId);
+
+    }
+
     public String sendGoOutAlarm(Long userId) {
 
-        UserData userData = userDataRepository.findById(userId).orElseThrow();
+        UserData userData = userDataRepository.findById(userId).orElseThrow(
+                () -> new CustomException(ErrorCode.NO_USER_DATA_REGISTERED));
 
         String finalMessage = "";
 
         /**
-         * 당일 온도 정리 - Weather Service
-         * {@link com.waither.notiservice.enums.Expressions} 참고
+         * 1. 기본 메세지 시작 형식
          */
         //TODO : 날씨 정보 가져오기 & 날씨별 멘트 정리
         String nickName = userData.getNickName();
         finalMessage += nickName + "님, 오늘은 ";
 
+
+        /**
+         * 2. 당일 온도 정리 - Weather Service
+         * {@link com.waither.notiservice.enums.Expressions} 참고
+         */
         double temperature = 10.8;
         if (userData.isUserAlert()) {
             //사용자 맞춤 알림이 on이라면 -> 계산 후 전용 정보 제공
-            UserMedian userMedian = userMedianRepository.findById(userId).orElseThrow();
+            UserMedian userMedian = userMedianRepository.findById(userId).orElseThrow(
+                    () ->  new CustomException(ErrorCode.NO_USER_MEDIAN_REGISTERED));
             finalMessage += TemperatureUtils.createUserDataMessage(userMedian, temperature);
         } else {
             //사용자 맞춤 알림이 off라면 -> 하루 평균 온도 정보 제공
             finalMessage += "평균 온도가 "+temperature+"도예요.";
         }
 
-        /**
-         * 미세먼지 정보 가져오기 - Weather Service
-         */
-        //TODO : 미세먼지 On/Off 확인, 멘트
-        finalMessage += " 미세먼지는 없습니다.";
+
+//        /**
+//         * 2. 미세먼지 정보 가져오기 - Weather Service
+//         */
+//        //TODO : 미세먼지 On/Off 확인, 멘트
+//        finalMessage += " 미세먼지는 없습니다.";
 
         /**
-         * 강수 정보 가져오기 - Weather Service
+         * 3. 강수 정보 가져오기 - Weather Service
          */
         //TODO : 강수량 확인, 멘트
         finalMessage += " 오후 6시부터 8시까지 120mm의 비가 올 예정입니다.";
 
         /**
-         * 꽃가로 정보 가져오기 - Weather Service
+         * 4. 꽃가루 정보 가져오기 - Weather Service
          */
         //TODO : 꽃가루 확인
         finalMessage += " 꽃가루는 없습니다. ";
